backend/requirements.txt
alembic==1.16.5
slowapi==0.1.9
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
bcrypt==4.1.2
click==8.3.0
ecdsa==0.19.1
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
httpx==0.27.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.12.0
pydantic-settings==2.9.1
pydantic_core==2.41.1
python-jose==3.5.0
python-multipart==0.0.20
passlib==1.7.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
email-validator==2.2.0
gunicorn==23.0.0
cryptography==44.0.3
reportlab==4.0.7
Pillow==10.1.0
python-dateutil==2.8.2